requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
django-environ==0.4.4
django-helusers==0.4.2
django==2.0.5
djangorestframework==3.8.2
drf-oidc-auth==0.9        # via django-helusers
future==0.16.0            # via pyjwkest
idna==2.6                 # via requests
psycopg2-binary==2.7.4
pycryptodomex==3.6.1      # via pyjwkest
pyjwkest==1.4.0           # via drf-oidc-auth
pytz==2018.4              # via django
raven==6.8.0
requests==2.18.4          # via django-helusers, pyjwkest
six==1.11.0               # via django-environ, pyjwkest
urllib3==1.22             # via requests
